feat: initial transforms API.
